backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
PyJWT==2.8.0